Move MockQuerier into separate module
