+++ packages/vm/src/testing/mod.rs
@@ -3,6 +3,7 @@
 mod calls;
 mod instance;
 mod mock;
+mod querier;
 mod storage;
 
 pub use calls::{handle, init, migrate, query};
@@ -11,9 +12,9 @@ pub use instance::{
     mock_instance_with_gas_limit, mock_instance_with_options, test_io, MockInstanceOptions,
 };
 pub use mock::{
-    mock_dependencies, mock_dependencies_with_balances, mock_env, MockApi, MockQuerier,
-    MOCK_CONTRACT_ADDR,
+    mock_dependencies, mock_dependencies_with_balances, mock_env, MockApi, MOCK_CONTRACT_ADDR,
 };
+pub use querier::MockQuerier;
 #[cfg(feature = "iterator")]
 pub use storage::MockIterator;
 pub use storage::MockStorage;
